backend/requirements.txt
Django==5.2.1
djangorestframework==3.16.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
djangorestframework-simplejwt==5.4.0
Pillow==11.2.1
django-filter==25.1
drf-spectacular==0.27.2
django-cors-headers==4.7.0
gunicorn==22.0.0
google-auth==2.40.3
requests==2.32.3
